EvalFiles

Description

Version : v1.0.0

Functions for evaluation and comparing of two files. Handles the report on equality.

Documentation

type FileContent = String

Type that represent file contents.

type FileName = String

Type that represent filename.

argFilesExist :: String -> String -> Bool -> IO ()

Prints error message to terminal if any of files does not exist. If both files exists evaluation is continued.

readFiles :: String -> String -> Bool -> IO ()

Read files from disc and calls the compare function and prints the report to terminal or file depending on argument toFile.

dateStr :: CalendarTime -> String

Returns a formated date and time string in format: YYYY-MM-DD HH:MM:SS

fileNameStr :: CalendarTime -> FileName -> FileName -> String

Returns name for report file formated as: YYYYMMDD_HHMMSS_filename1_ext_filename2_ext.txt Characters: .(dot), /(slash) and (backslash) is replaced with underscore.

padNoStr :: Int -> String

Pads an integer with a 0 (zero) if smaller than 10. Returns a string.

monthNoStr :: Month -> String

Returns a paded string that represent the month number. If month number smaller than 10 then a 0 (zero) is paded before the month number.

parseFiles :: [FileContent] -> [FileContent] -> [String] -> Int -> Int -> String

Function takes contents for two files (splited in lines) to be compared, the default array of Strings that is the formated Strings stating the difference, usualy an empty array, and an integer counting up the number of rows parsed and an integer counting up the number of different rows. Both usualy zero. Returning the formated report String from the parsing of the content.

sumParse :: [String] -> Int -> Int -> String

The final text from the parsing of the files. Number of rows searched and number of rows that differs.